+++ cloudsql_dumps_bucket.tf
@@ -3,23 +3,27 @@
 # --------------------------------
 # Retrieve the CloudSQL instance.
 data "google_sql_database_instance" "cloudsql_instance" {
-  count   = var.create_clousql_dumps_bucket ? 1 : 0
+  count   = var.create_cloudsql_dumps_bucket ? 1 : 0
   name    = var.cloudsql_instance_name
   project = var.project_id
 }
 
 # Create a random suffix for the bucket name.
 resource "random_id" "cloudsql_dumps_bucket_name_suffix" {
-  count       = var.create_clousql_dumps_bucket ? 1 : 0
+  count       = var.create_cloudsql_dumps_bucket ? 1 : 0
   byte_length = 4
 }
 
 locals {
-  cloudsql_dumps_bucket_name = var.create_clousql_dumps_bucket ? "${var.project_id}-cloudsql-dumps-${random_id.cloudsql_dumps_bucket_name_suffix[0].hex}" : null
+  cloudsql_dumps_bucket_name = var.create_cloudsql_dumps_bucket ? (
+    var.cloudsql_dumps_bucket_name != "" ? "${var.cloudsql_dumps_bucket_name}-${random_id.cloudsql_dumps_bucket_name_suffix[0].hex}" : (
+      var.cloudsql_dumps_environment != "" ? "${var.project_id}-cloudsql-dumps-${var.cloudsql_dumps_environment}-${random_id.cloudsql_dumps_bucket_name_suffix[0].hex}" : "${var.project_id}-cloudsql-dumps-${random_id.cloudsql_dumps_bucket_name_suffix[0].hex}"
+    )
+  ) : null
 }
 
 resource "google_storage_bucket" "cloudsql_dumps" {
-  count         = var.create_clousql_dumps_bucket ? 1 : 0
+  count         = var.create_cloudsql_dumps_bucket ? 1 : 0
   name          = local.cloudsql_dumps_bucket_name
   location      = var.region
   storage_class = "NEARLINE"
@@ -35,7 +39,7 @@ resource "google_storage_bucket" "cloudsql_dumps" {
 
   lifecycle_rule {
     action {
-      type = "SetStorageClass"
+      type          = "SetStorageClass"
       storage_class = "COLDLINE"
     }
     condition {
@@ -53,7 +57,7 @@ resource "google_storage_bucket" "cloudsql_dumps" {
 # Cloud SQL console, but it must be done manually if you want to use the
 # `gcloud sql export` command.
 resource "google_storage_bucket_iam_member" "cloudsql_dumps_bucket_writer" {
-  count  = var.create_clousql_dumps_bucket ? 1 : 0
+  count  = var.create_cloudsql_dumps_bucket ? 1 : 0
   bucket = google_storage_bucket.cloudsql_dumps[0].name
   role   = "roles/storage.legacyBucketWriter"
   member = "serviceAccount:${data.google_sql_database_instance.cloudsql_instance[0].service_account_email_address}"
@@ -64,7 +68,7 @@ resource "google_storage_bucket_iam_member" "cloudsql_dumps_bucket_writer" {
 # as the index 0, and the tag value shortname as the index 1.
 # The friendly name is in the form <TAG_KEY_SHORTNAME>/<TAG_VALUE_SHORTNAME>
 data "google_tags_tag_key" "tag_keys" {
-  for_each   = var.create_clousql_dumps_bucket ? toset(var.global_tags) : []
+  for_each   = var.create_cloudsql_dumps_bucket ? toset(var.global_tags) : []
   parent     = "projects/${var.project_id}"
   short_name = split("/", each.value)[0]
 }
@@ -72,14 +76,14 @@ data "google_tags_tag_key" "tag_keys" {
 # To bind a tag to a resource, we need to know the tag value ID (something as
 # "tagValues/281483307043046"), that we can retrieve from this data source.
 data "google_tags_tag_value" "tag_values" {
-  for_each   = var.create_clousql_dumps_bucket ? toset(var.global_tags) : []
+  for_each   = var.create_cloudsql_dumps_bucket ? toset(var.global_tags) : []
   parent     = data.google_tags_tag_key.tag_keys[each.value].id
   short_name = split("/", each.value)[1]
 }
 
 # Bind tags to buckets.
 resource "google_tags_location_tag_binding" "binding" {
-  for_each  = var.create_clousql_dumps_bucket ? toset(var.global_tags) : []
+  for_each  = var.create_cloudsql_dumps_bucket ? toset(var.global_tags) : []
   parent    = "//storage.googleapis.com/projects/_/buckets/${google_storage_bucket.cloudsql_dumps[0].name}"
   location  = google_storage_bucket.cloudsql_dumps[0].location
   tag_value = data.google_tags_tag_value.tag_values[each.value].id